Commit the blended character animation work.  It was an interestin day.  I started out by trying to make more compelling design models.  The first attempt was a butterfly.  I used a shape key to animate the transition from wings open to wings closed.  After testing it out in GreenRoom, I realized that the wings were not rotating around a single point at the root of the wing.  Instead the wingtips were transitioning along a line from the Basis shape key to the WingsUp shape key.  Bad, bad, bad.
Therefore, I went ahead and tried to get bones working within GreenRoom.  I was initially successful by adding bones to the butterfly model, and animating it with a slider in the GUI.  However, to do that, I had to hard code the notion of rotating the bones 80 degrees about the root, and the fact the left and right wings move together.  To make this work with standard design models, we'd have to create a way to store such data.  So, instead of doing that, I spent time trying to figure out how Blender and Unity control character animations.  The end result (for now) is faking the storage of poses within single frame animations.  I can call Animation.CrossFade("Wings Up", 1.0) to cause Unity to 'mix' the different positions.  Pretty cool.