Finish the simple example of running CuraEngine on the server.  Next steps:
- implement unique ID files using an uploads table in the database
- verify that Gcode is being gzipped when it is being sent back to the browser

Ulitmately, this utility only demonstrates that we can run CuraEngine on the server.  I don't expect us to use a script like this in the prototype.  Instead, we need to implement a driver application for Windows and Mac OS X that will connect to the user's 3D printer, wait for incoming print jobs, then print them.