Modify BlitItObject, BlitItMeshRenderer, and Attachment's so they can be copied.
This modifies their initialization routines to behave appropriately if the object is copied.  For example, BlitItObject will check the 'blitItObject' reference of attachments, and only destroy them if it matches their 'this' pointer (in other words, the attachment is owned by the BlitItObject).

Also, BlitItMeshRenderer's initialization routine has changed such that it doesn't destroy pre-existing MeshRenderer, MeshFilter, or MeshCollider.  This is necessary because those would legitimately be provided if the object were instantiated from a prefab, or copied via the 'Duplicate' menu in the Unity editor.