Make the blititd project a sub-project of Unity-iPhone.
To set this up, first drag and drop the blititd project into the project panel.  That makes it a subproject.  Then, for your intended build target, you have to go into the Build Phases and add a target dependency on blititeditor_iPad (hit plus, then select blititeditor_iPad).  Then, drag and drop blititeditor_iPad.a from blititd's Products folder in the project panel into the 'link with libraries' build phase.

I know it sounds convoluted, but conceptually it isn't.  It's simply that an Xcode project has a dependency on the build product of another.  This is the way Apple does it with their own projects.