Imported standard image effect assets; background now blurs when main menu panel is opened.
Menu is still not entirely modal; clicking elsewhere on the screen while the menu is open will close the menu but also perform the action of whatever UI element was clicked.

Also introduced a bug where clicking on an item in the object store will immediately create the object in the center of the workplane instead of displaying the object info panel.