Working on shell generator. Current script is very basic and has several issues that need to be addressed. Concave spaces are potentially problematic depending on the extent of the expansion and need to be addressed in some way.
Also, the way vertices are moved needs some work; currently, the script gets all vertices with the same position and gathers the normalized sum of their normals. This works in most cases, but causes some inconsistent scaling in certain situations. For example, a vertex that is on an edge won't be moved as much as a vertex in a flat area.