Added the following tables to the databse:
- checking_accounts
- transactions
- designer_disbursement_transactions

The checking accounts table stores the routing and account numbers of the designers.  Payments are made at regular intervals (once per month) to the designer, compensating them for sales of their designs.

The transactions table is the primary way to track all end-user purchases.  It contains a column that indicates whether the designer has been compensated for each transaction.  If the developer has not been compensated for that transaction, then that transaction will be picked up in the next designer disbursement.

designer_disbursement_transactions is the record of all disbursements to designers.  This is our record that indicates that a particular designer has been paid.