Fix for Bug 273 - BlitItDataStore::AddOwnedObject - race condition
https://bugs.blitit.com/show_bug.cgi?id=273

It turns out that the check as to whether the user already owns the object was totally unnecessary.  The owned_objects table uses both owner_id and owned_object_id together as the primary key.  Therefore, if our code ever attempts to add a duplicate ownership record, the query will fail.