Fix for Bug 406 - Websites need a favicon
https://bugs.blitit.com/show_bug.cgi?id=406

Add the favicons to the designer portal, too.  This kind of sucks, actually.  These files are duplicated between blitit.com and designer.blitit.com.  This is caused by the fact that 1) favicon's are ideally placed in the root of the domain (i.e. DocumentRoot) and 2) subversion doesn't support svn:externals where files are "dumped into" a directory that already exists in the repo.