Work towards Bug 397 - Don't store checking account information in the clear.
Since, for security reasons, we would not be exposing the checking account number in the web interface, we can't display it in the 'Edit Account Details' form.  Since we can't show the checking account number here, it doesn't make much sense to show *any* of the checking account information.  For this to make sense, you'd have to argue that changing the name or birth date associated with an account is by far the most common scenario rather than simply replacing the checking account with an entirely new one.  I don't think that's the case.