Update.
Well... after working on the triangle-based version of this for a while, it's apparent that it isn't going to work.  The triangles do not express edges that collapse to zero when a hole poly intersects with a container poly.  So... we have to go with a "bridge edge" implementation.  A bridge edge implementation will create bridge edges that shoot across each miter normal, and thereby indicate precisely when intersections and collapses occur.